Добавлены сущности 29.10.2018 22:21

+++ src/main/java/com/railwaycompany/dao/api/GenericDao.java
@@ -7,7 +7,7 @@ public interface GenericDao<E extends Serializable> {
 
     void create(E entity);
 
-    E read(int key);
+    E read(long key);
 
     List<E> readAll();
